+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/types/filters/BlendComposite.java
@@ -172,6 +172,7 @@ public final class BlendComposite implements Composite {
 
     /**
      * Gets alpha.
+     *
      * @return Alpha
      */
     public float getAlpha() {