+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/types/filters/COLORMATRIXFILTER.java
@@ -119,5 +119,4 @@ public class COLORMATRIXFILTER extends FILTER {
         return Arrays.equals(this.matrix, other.matrix);
     }
 
-    
 }