+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/types/filters/CONVOLUTIONFILTER.java
@@ -194,6 +194,5 @@ public class CONVOLUTIONFILTER extends FILTER {
         }
         return Objects.equals(this.defaultColor, other.defaultColor);
     }
-    
-    
+
 }